fix tests for v3 schema updates
