+++ grype/presenter/json/test-fixtures/snapshot/TestJsonDirsPresenter.golden
@@ -3,7 +3,9 @@
   {
    "vulnerability": {
     "id": "CVE-1999-0001",
+    "dataSource": "",
     "severity": "Low",
+    "urls": [],
     "description": "1999-01 description",
     "cvss": [
      {
@@ -15,11 +17,18 @@
       "vendorMetadata": {}
      }
     ],
-    "fixedInVersion": "the-next-version"
+    "fix": {
+     "versions": [
+      "the-next-version"
+     ],
+     "state": ""
+    },
+    "advisories": []
    },
+   "relatedVulnerabilities": [],
    "matchDetails": {
     "matcher": "dpkg-matcher",
-    "searchKey": {
+    "searchedBy": {
      "distro": {
       "type": "ubuntu",
       "version": "20.04"
@@ -39,7 +48,7 @@
      }
     ],
     "language": "",
-    "licenses": null,
+    "licenses": [],
     "cpes": [],
     "purl": "",
     "metadata": {
@@ -50,7 +59,9 @@
   {
    "vulnerability": {
     "id": "CVE-1999-0002",
+    "dataSource": "",
     "severity": "Critical",
+    "urls": [],
     "description": "1999-02 description",
     "cvss": [
      {
@@ -66,11 +77,17 @@
        "Status": "verified"
       }
      }
-    ]
+    ],
+    "fix": {
+     "versions": [],
+     "state": ""
+    },
+    "advisories": []
    },
+   "relatedVulnerabilities": [],
    "matchDetails": {
     "matcher": "dpkg-matcher",
-    "searchKey": {
+    "searchedBy": {
      "cpe": "somecpe"
     },
     "matchedOn": {
@@ -87,7 +104,7 @@
      }
     ],
     "language": "",
-    "licenses": null,
+    "licenses": [],
     "cpes": [],
     "purl": "",
     "metadata": {
@@ -98,14 +115,23 @@
   {
    "vulnerability": {
     "id": "CVE-1999-0003",
+    "dataSource": "",
     "severity": "High",
+    "urls": [],
     "description": "1999-03 description",
     "cvss": [],
-    "fixedInVersion": "the-other-next-version"
+    "fix": {
+     "versions": [
+      "the-other-next-version"
+     ],
+     "state": ""
+    },
+    "advisories": []
    },
+   "relatedVulnerabilities": [],
    "matchDetails": {
     "matcher": "dpkg-matcher",
-    "searchKey": {
+    "searchedBy": {
      "language": "java"
     },
     "matchedOn": {
@@ -122,7 +148,7 @@
      }
     ],
     "language": "",
-    "licenses": null,
+    "licenses": [],
     "cpes": [],
     "purl": "",
     "metadata": {
