+++ grype/presenter/json/test-fixtures/snapshot/TestJsonImgsPresenter.golden
@@ -3,7 +3,9 @@
   {
    "vulnerability": {
     "id": "CVE-1999-0001",
+    "dataSource": "",
     "severity": "Low",
+    "urls": [],
     "description": "1999-01 description",
     "cvss": [
      {
@@ -15,11 +17,18 @@
       "vendorMetadata": {}
      }
     ],
-    "fixedInVersion": "the-next-version"
+    "fix": {
+     "versions": [
+      "the-next-versions"
+     ],
+     "state": ""
+    },
+    "advisories": []
    },
+   "relatedVulnerabilities": [],
    "matchDetails": {
     "matcher": "dpkg-matcher",
-    "searchKey": {
+    "searchedBy": {
      "distro": {
       "type": "ubuntu",
       "version": "20.04"
@@ -40,7 +49,7 @@
      }
     ],
     "language": "",
-    "licenses": null,
+    "licenses": [],
     "cpes": [
      "cpe:2.3:a:anchore:engine:0.9.2:*:*:python:*:*:*:*"
     ],
@@ -53,7 +62,9 @@
   {
    "vulnerability": {
     "id": "CVE-1999-0002",
+    "dataSource": "",
     "severity": "Critical",
+    "urls": [],
     "description": "1999-02 description",
     "cvss": [
      {
@@ -69,11 +80,17 @@
        "Status": "verified"
       }
      }
-    ]
+    ],
+    "fix": {
+     "versions": [],
+     "state": ""
+    },
+    "advisories": []
    },
+   "relatedVulnerabilities": [],
    "matchDetails": {
     "matcher": "dpkg-matcher",
-    "searchKey": {
+    "searchedBy": {
      "cpe": "somecpe"
     },
     "matchedOn": {
@@ -91,7 +108,7 @@
      }
     ],
     "language": "",
-    "licenses": null,
+    "licenses": [],
     "cpes": [
      "cpe:2.3:a:anchore:engine:0.9.2:*:*:python:*:*:*:*"
     ],
@@ -104,14 +121,23 @@
   {
    "vulnerability": {
     "id": "CVE-1999-0003",
+    "dataSource": "",
     "severity": "High",
+    "urls": [],
     "description": "1999-03 description",
     "cvss": [],
-    "fixedInVersion": "the-other-next-version"
+    "fix": {
+     "versions": [
+      "the-other-next-version"
+     ],
+     "state": ""
+    },
+    "advisories": []
    },
+   "relatedVulnerabilities": [],
    "matchDetails": {
     "matcher": "dpkg-matcher",
-    "searchKey": {
+    "searchedBy": {
      "language": "java"
     },
     "matchedOn": {
@@ -129,7 +155,7 @@
      }
     ],
     "language": "",
-    "licenses": null,
+    "licenses": [],
     "cpes": [
      "cpe:2.3:a:anchore:engine:0.9.2:*:*:python:*:*:*:*"
     ],
