+++ grype/presenter/models/vulnerability_metadata.go
@@ -7,22 +7,9 @@ type VulnerabilityMetadata struct {
 	DataSource  string   `json:"dataSource"`
 	Namespace   string   `json:"namespace,omitempty"`
 	Severity    string   `json:"severity,omitempty"`
-	URLs        []string `json:"urls,omitempty"`
+	URLs        []string `json:"urls"`
 	Description string   `json:"description,omitempty"`
-	Cvss        []Cvss   `json:"cvss,omitempty"`
-}
-
-type Cvss struct {
-	Version        string      `json:"version"`
-	Vector         string      `json:"vector"`
-	Metrics        CvssMetrics `json:"metrics"`
-	VendorMetadata interface{} `json:"vendorMetadata"`
-}
-
-type CvssMetrics struct {
-	BaseScore           float64  `json:"baseScore"`
-	ExploitabilityScore *float64 `json:"exploitabilityScore,omitempty"`
-	ImpactScore         *float64 `json:"impactScore,omitempty"`
+	Cvss        []Cvss   `json:"cvss"`
 }
 
 func NewVulnerabilityMetadata(id, namespace string, metadata *vulnerability.Metadata) VulnerabilityMetadata {
@@ -33,22 +20,9 @@ func NewVulnerabilityMetadata(id, namespace string, metadata *vulnerability.Meta
 		}
 	}
 
-	cvss := make([]Cvss, 0)
-	for _, score := range metadata.Cvss {
-		vendorMetadata := score.VendorMetadata
-		if vendorMetadata == nil {
-			vendorMetadata = make(map[string]interface{})
-		}
-		cvss = append(cvss, Cvss{
-			Version: score.Version,
-			Vector:  score.Vector,
-			Metrics: CvssMetrics{
-				BaseScore:           score.Metrics.BaseScore,
-				ExploitabilityScore: score.Metrics.ExploitabilityScore,
-				ImpactScore:         score.Metrics.ImpactScore,
-			},
-			VendorMetadata: vendorMetadata,
-		})
+	urls := metadata.URLs
+	if urls == nil {
+		urls = make([]string, 0)
 	}
 
 	return VulnerabilityMetadata{
@@ -56,8 +30,8 @@ func NewVulnerabilityMetadata(id, namespace string, metadata *vulnerability.Meta
 		DataSource:  metadata.DataSource,
 		Namespace:   metadata.Namespace,
 		Severity:    metadata.Severity,
-		URLs:        metadata.URLs,
+		URLs:        urls,
 		Description: metadata.Description,
-		Cvss:        cvss,
+		Cvss:        NewCVSS(metadata),
 	}
 }
