+++ grype/presenter/table/presenter_test.go
@@ -37,8 +37,8 @@ func TestTablePresenter(t *testing.T) {
 	var match1 = match.Match{
 		Type: match.ExactDirectMatch,
 		Vulnerability: vulnerability.Vulnerability{
-			ID:           "CVE-1999-0001",
-			RecordSource: "source-1",
+			ID:        "CVE-1999-0001",
+			Namespace: "source-1",
 		},
 		Package: pkg1,
 		Matcher: match.DpkgMatcher,
@@ -47,9 +47,13 @@ func TestTablePresenter(t *testing.T) {
 	var match2 = match.Match{
 		Type: match.ExactIndirectMatch,
 		Vulnerability: vulnerability.Vulnerability{
-			ID:             "CVE-1999-0002",
-			RecordSource:   "source-2",
-			FixedInVersion: "the-next-version",
+			ID:        "CVE-1999-0002",
+			Namespace: "source-2",
+			Fix: vulnerability.Fix{
+				Versions: []string{
+					"the-next-version",
+				},
+			},
 		},
 		Package: pkg2,
 		Matcher: match.DpkgMatcher,
