+++ grype/vulnerability/store_adapter_test.go
@@ -36,12 +36,14 @@ func TestGetByDistro(t *testing.T) {
 			ID:         "CVE-2014-fake-1",
 			Namespace:  "debian:8",
 			CPEs:       []syftPkg.CPE{},
+			Advisories: []Advisory{},
 		},
 		{
 			Constraint: version.MustGetConstraint("< 2013.0.2-1", version.DebFormat),
 			ID:         "CVE-2013-fake-2",
 			Namespace:  "debian:8",
 			CPEs:       []syftPkg.CPE{},
+			Advisories: []Advisory{},
 		},
 	}
 
@@ -80,7 +82,8 @@ func TestGetByCPE(t *testing.T) {
 					CPEs: []syftPkg.CPE{
 						must(syftPkg.NewCPE("cpe:2.3:*:activerecord:activerecord:*:*:something:*:*:ruby:*:*")),
 					},
-					Namespace: "nvd",
+					Namespace:  "nvd",
+					Advisories: []Advisory{},
 				},
 			},
 		},
@@ -95,7 +98,8 @@ func TestGetByCPE(t *testing.T) {
 					CPEs: []syftPkg.CPE{
 						must(syftPkg.NewCPE("cpe:2.3:*:activerecord:activerecord:*:*:*:*:*:rails:*:*")),
 					},
-					Namespace: "nvd",
+					Namespace:  "nvd",
+					Advisories: []Advisory{},
 				},
 				{
 					Constraint: version.MustGetConstraint("< 3.7.4", version.UnknownFormat),
@@ -103,7 +107,8 @@ func TestGetByCPE(t *testing.T) {
 					CPEs: []syftPkg.CPE{
 						must(syftPkg.NewCPE("cpe:2.3:*:activerecord:activerecord:*:*:something:*:*:ruby:*:*")),
 					},
-					Namespace: "nvd",
+					Namespace:  "nvd",
+					Advisories: []Advisory{},
 				},
 			},
 		},
